# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "srvcs-range",
    "description": "",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/": {
      "get": {
        "tags": [],
        "summary": "`GET /` — service identity (srvcs service standard).",
        "operationId": "index",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Info"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [],
        "summary": "`POST /` — compute the range (max - min) of a list of integers.",
        "description": "This service does no arithmetic of its own. It asks `srvcs-sortascending`\nfor the sorted list, then asks `srvcs-subtract` for\n`sorted[last] - sorted[0]`. The empty list is rejected with `422`. Invalid\nelements are rejected by the leaf dependencies and the resulting `422` is\nforwarded unchanged.",
        "operationId": "evaluate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RangeResponse"
                }
              }
            }
          },
          "422": {
            "description": "empty list, or an element is not a valid integer (forwarded)"
          },
          "500": {
            "description": "a dependency returned an unusable response"
          },
          "503": {
            "description": "a dependency is unavailable"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "EvalRequest": {
        "type": "object",
        "required": [
          "values"
        ],
        "properties": {
          "values": {
            "type": "object",
            "description": "The list of integers to take the range of. An empty list is rejected."
          }
        }
      },
      "Info": {
        "type": "object",
        "required": [
          "service",
          "concern",
          "depends_on"
        ],
        "properties": {
          "concern": {
            "type": "string"
          },
          "depends_on": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "service": {
            "type": "string"
          }
        }
      },
      "RangeResponse": {
        "type": "object",
        "required": [
          "values",
          "result"
        ],
        "properties": {
          "result": {
            "type": "integer",
            "format": "int64"
          },
          "values": {
            "type": "object"
          }
        }
      }
    }
  }
}